requirements.txt
django-environ==0.12.0
Django==5.1.7
environ==1.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
psycopg==3.2.9
typing_extensions==4.13.2
psycopg2-binary==2.9.10